Updated Task 0 function code - Changed numNodes from size_t to unsigned

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -6,9 +6,10 @@
  * @h: Pointer to head node
  * Return: Number of nodes
  */
+
 size_t print_listint(const listint_t *h)
 {
-	size_t numNodes = 0;
+	unsigned numNodes = 0;
 	const listint_t *current = h;
 
 	if (h == NULL)
